Ny design av notat

+++ bidrag-dokumentmal/app/utils/visningsnavn.ts
@@ -62,7 +62,7 @@ export const formatterProsent = (
   value: number | string | undefined,
 ): string => {
   if (!value) return "0%";
-  const asNumber = typeof value == "string" ? parseInt(value) : value;
+  const asNumber = typeof value == "string" ? parseFloat(value) : value;
   const percentageAsFraction = asNumber / 100;
   return percentageAsFraction.toLocaleString("nb-NO", {
     style: "percent",
